+++ src/analysis/occurrences.ml
@@ -195,7 +195,10 @@ let get_external_locs ~(config : Mconfig.t) ~current_buffer_path uid :
       in
       Option.map external_locs ~f:(fun (index, locs) ->
           let stats = Stat_check.create ~cache_size:128 index in
-          ( Occurrence_set.of_filtered_lid_set locs ~f:(fun lid ->
+          (* Note: [get_outdated_files] must be evaluated after the [Stat_check.check stats ~file...],
+             as the latter one populates the cache inside [stats] *)
+          let occurrences =
+            Occurrence_set.of_filtered_lid_set locs ~f:(fun lid ->
                 let { Location.loc; _ } = Index_format.Lid.to_lid lid in
                 (* We ignore external results that concern the current buffer *)
                 let file_rel_to_root =
@@ -232,8 +235,9 @@ let get_external_locs ~(config : Mconfig.t) ~current_buffer_path uid :
                     | false -> Stale
                   in
                   Some staleness
-                end),
-            Stat_check.get_outdated_files stats )))
+                end)
+          in
+          (occurrences, Stat_check.get_outdated_files stats)))
 
 let lookup_related_uids_in_indexes ~(config : Mconfig.t) uid =
   let title = "lookup_related_uids_in_indexes" in
@@ -358,6 +362,11 @@ let locs_of ~config ~env ~typer_result ~pos ~scope path =
           (Occurrence_set.union acc_locs locs, String.Set.union acc_files files))
         external_occurrences
     in
+    if not (String.Set.is_empty out_of_sync_files) then
+      log ~title:"locs_of"
+        "Occurrences may be incomplete: some source files are out-of-sync with \
+         the index: %s"
+        (String.concat ~sep:", " (String.Set.to_list out_of_sync_files));
     let occurrences =
       Occurrence_set.union buffer_occurrences external_occurrences
     in